mtd: nand: add a generic data scrambler infrastructure
This patch introduce a new layer in the NAND framework to support both HW
and SW data scramblers.

The scrambling process is required on some MLC/TLC NAND chips which do not
support large islands of identical patterns.

The scrambler layer defines a nand_scrambler_ctrl struct which is intended
to be used by NAND core functions or NAND drivers to scramble/descramble
data stored on NAND chips.

The implementation can implement any of these functions:
- ->config(): prepare a random transfer to/from the NAND chip
- ->write_buf(): randomize and write data to the NAND chip
- ->read_buf(): read and derandomize data from the NAND chip

The ->read_buf()/->write_buf() functions are always called after a
->config() call.
The config call specify the page, the column within the page and the
action that will take place after the config (either read or write).
If column is set to -1, the randomizer is disabled.
If page is set to -1, we keep working on the same page.

The randomizer layer provides helper functions that choose whether the
scrambler or the chip ->read_buf()/->write_buf() methods should be used.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>