mtd: nand: add infrastructure for per-partition ECC config
Some SoCs require a different ECC config for their first stage bootloader,
but we cannot apply the same config on the whole flash, or we might have
to live an unsuitable ECC config (often weaker than what is required by
the NAND chip).

This patch makes use of the mtd_part_ops infrastructure to let the NAND
controller adjust the ECC config for each partition.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>