mtd: allow device-specific partition handling
Currently, only the mtdpart implementation is allowed to expose MTD device
partitions, but some devices might need specific handling for special
partitions.
This is particularly true for systems where the NAND device is used as a
boot media. On such systems, the ROM code only support a subset of ECC
modes, and this subset may not match the chip requirements. In this
particular case we only have two options:
- force the whole chip to use one of the ECC mode supported by the ROM code
- allow for per-partition ECC setting
The first solution should be avoided if the ECC setting used by the ROM
code are weaker than the NAND requirements, hence

Add a generic way to overload the default partition handling so that
MTD drivers (or sub-frameworks) can implement their own methods.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>